refactor(prd): address review feedback on stress-test UI (#561)
- Add @rate_limit_standard() to the stress-test SSE endpoint (it runs
  recursive LLM decomposition — throttle burst abuse), matching the rest
  of prd_v2. (coderabbit Major / claude High)
- Move SSE event payload types to web-ui/src/types/prd.ts (re-exported
  from types/index.ts) per the repo's type-location guideline; the hook
  imports them. (coderabbit Major)
- Guard useStressTestStream.start() when workspacePath is null — fail
  fast to the error state instead of hanging in 'streaming'. (coderabbit)
- Drop the redundant ScrollArea wrapper in StressTestModal so the
  auto-scroll ref and the visible scrollbar are the same element.
  (claude Medium)
- Tighten the disconnect-abort test assertion to <= 2 calls (this
  fixture's full run is 4). (coderabbit)
